Conflicts:
	ZaiLuShang2/ZaiLuShang2.xcodeproj/project.pbxproj

ZaiLuShang2/ZaiLuShang2.xcodeproj/project.xcworkspace/xcuserdata/qianfen
g.xcuserdatad/UserInterfaceState.xcuserstate

ZaiLuShang2/ZaiLuShang2.xcodeproj/xcuserdata/qianfeng.xcuserdatad/xcdebu
gger/Breakpoints_v2.xcbkptlist

+++ ZaiLuShang2/ZaiLuShangClass/Search(搜索)/controller/SearchVC.m
@@ -116,7 +116,7 @@
 
 -(NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
-    int num =[[_dataArray[section] listArray] count];
+    NSUInteger num =[[_dataArray[section] listArray] count];
     if (num>1) {
         return num*0.5;
     }
